BACKUP/requirements.txt
pandas
matplotlib
numpy
torch
torchvision
pyyaml